Also reorganized tests and App.tsx
